rj.gd {rj.gd}	R Documentation

Create a new RJ graphics device

Description

.rj.GD initializes a new graphics device. If supported, the new device is shown in a new window/page in the client application.

Usage

.rj.GD(name = "rj.gd", width = 7, height = 7, size.unit = "in",
    xpinch = NA, ypinch = NA, canvas = "white",
    pointsize = 12, gamma = 1.0)

Arguments

name
Name of the device (unused)

width, height
Initial size (number) of the graphic window, for horizontal and vertical orientation. The unit of the values are inch, if not otherwise specified by the argument size.unit. The default size is 7 x 7 inch.

size.unit
The unit (known character) of the size values in width and height. Supported units are inch ("in") and pixels ("px"). The default unit is inch.

xpinch, ypinch
The resolution of the device (number), pixel per inch (ppi), for horizontal and vertical orientation. By default the values are specified by the client application.

canvas
The color (R color) of the canvas which is visible when the background color is transparent. The default color is "white".

pointsize
The default pointsize (number) of plotted text in big points (1/72 inch). Values are rounded to the nearest integer. The default text size is 12 pt.

gamma
Gamma correction factor as number (not the gamma value itself). Colors in R are defined for the color space sRGB. If your monitor does not conform to sRGB, you might be able to improve things by tweaking this parameter to apply additional gamma correction to the RGB channels. The default is no additional gamma correction (1.0).

Value

Currently the return value is NULL.

Examples

## Not run: 
.rj.GD()

## End(Not run)

[Package rj.gd version 2.1.0-2 Index]